Fixed a bug related to the use of 'use client' in page.tsx - created new components. Removed all uses of console.log, added new types
