+++ next.config.mjs
@@ -5,6 +5,4 @@ const nextConfig = {
   },
 };
 
-const NEXTAUTH_URL='http://localhost:3001'
-
 export default nextConfig;
